Update system check and relative image path
Added a system check in Main.groovy to ensure that the program is running on a Raspberry Pi and with root access. Also, switched an absolute image path to a relative one in README.md to fix an image display issue. In addition, updated .gitignore to ignore the /target/ directory.